Started to move some mocks to typescript
