--- __mocks__/file-mock.js
@@ -1,1 +1,0 @@
-module.exports = 'test-file-stub';
